pfs-analysis/requirements.txt
streamlit
pandas
altair
plotly
psycopg2-binary
python-dotenv
requests
